boundary.efp {strucchange}	R Documentation

Boundary for Empirical Fluctuation Processes

Description

Computes boundary for an object of class "efp"

Usage


## S3 method for class 'efp':
boundary(x, alpha = 0.05, alt.boundary = FALSE,
   functional = "max", ...)

Arguments

x 	 an object of class "efp".

alpha 	 numeric from interval (0,1) indicating the confidence level for which the boundary of the corresponding test will be computed.

alt.boundary 	 logical. If set to TRUE alternative boundaries (instead of the standard linear boundaries) will be computed (for Brownian bridge type processes only).

functional 	 indicates which functional should be applied to the empirical fluctuation process. See also plot.efp.

... 	 currently not used.

Value

an object of class "ts" with the same time properties as the process in x

See Also

efp, plot.efp

Examples


## Load dataset "nhtemp" with average yearly temperatures in New Haven
data(nhtemp)
## plot the data
plot(nhtemp)

## test the model null hypothesis that the average temperature remains constant
## over the years
## compute OLS-CUSUM fluctuation process
temp.cus <- efp(nhtemp ~ 1, type = "OLS-CUSUM")
## plot the process without boundaries
plot(temp.cus, alpha = 0.01, boundary = FALSE)
## add the boundaries in another colour
bound <- boundary(temp.cus, alpha = 0.01)
lines(bound, col=4)
lines(-bound, col=4)

[Package Contents]