gefp {strucchange}	R Documentation

Generalized Empirical M-Fluctuation Processes

Description

Computes an empirical M-fluctuation process from the scores of a fitted model.

Usage

gefp(..., fit = glm, scores = estfun, vcov = NULL,
  decorrelate = TRUE, sandwich = TRUE, order.by = NULL,
  fitArgs = NULL, parm = NULL, data = list())

Arguments

...
specification of some model which is passed together with data to the fit function: fm <- fit(..., data = data). If fit is set to NULL the first argument ... is assumed to be already the fitted model fm (all other arguments in ... are ignored and a warning is issued in this case).

fit
a model fitting function, typically lm, glm or rlm.

scores
a function which extracts the scores or estimating function from the fitted object: scores(fm).

vcov
a function to extract the covariance matrix for the coefficients of the fitted model: vcov(fm, order.by = order.by, data = data).

decorrelate
logical. Should the process be decorrelated?

sandwich
logical. Is the function vcov the full sandwich estimator or only the meat?

order.by
Either a vector z or a formula with a single explanatory variable like ~ z. The observations in the model are ordered by the size of z. If set to NULL (the default) the observations are assumed to be ordered (e.g., a time series).

fitArgs
List of additional arguments which could be passed to the fit function. Usually, this is not needed and ... will be sufficient to pass arguments to fit.

parm
integer or character specifying the component of the estimating functions which should be used (by default all components are used).

data
an optional data frame containing the variables in the ... specification and the order.by model. By default the variables are taken from the environment which gefp is called from.

Value

gefp returns a list of class "gefp" with components including:

process
the fitted empirical fluctuation process of class "zoo",

nreg
the number of regressors,

nobs
the number of observations,

fit
the fit function used,

scores
the scores function used,

fitted.model
the fitted model.

References

Zeileis A. (2005), A Unified Approach to Structural Change Tests Based on ML Scores, F Statistics, and OLS Residuals. Econometric Reviews, 24, 445–466. doi:10.1080/07474930500406053.

Zeileis A. (2006), Implementing a Class of Structural Change Tests: An Econometric Computing Approach. Computational Statistics & Data Analysis, 50, 2987–3008. doi:10.1016/j.csda.2005.07.001.

Zeileis A., Hornik K. (2007), Generalized M-Fluctuation Tests for Parameter Instability, Statistica Neerlandica, 61, 488–508. doi:10.1111/j.1467-9574.2007.00371.x.

Zeileis A., Shah A., Patnaik I. (2010), Testing, Monitoring, and Dating Structural Changes in Exchange Rate Regimes, Computational Statistics and Data Analysis, 54(6), 1696–1706. doi:10.1016/j.csda.2009.12.005.

See Also

efp, efpFunctional

Examples

data("BostonHomicide")
gcus <- gefp(homicides ~ 1, family = poisson, vcov = kernHAC,
	     data = BostonHomicide)
plot(gcus, aggregate = FALSE)	 
gcus
sctest(gcus)

[Package strucchange version 1.5-3 Index]