plot.mefp {strucchange}	R Documentation

Plot Methods for mefp Objects

Description

This is a method of the generic plot function for for "mefp" objects as returned by mefp or monitor. It plots the emprical fluctuation process (or a functional therof) as a time series plot, and includes boundaries corresponding to the significance level of the monitoring procedure.

Usage


## S3 method for class 'mefp':
plot(x, boundary = TRUE, functional = "max", main = NULL,
    ylab = "empirical fluctuation process", ylim = NULL, ...)

Arguments

x 	 an object of class "mefp".

boundary 	 if FALSE, plotting of boundaries is suppressed.

functional 	 indicates which functional should be applied to a multivariate empirical process. If set to NULL all dimensions of the process (one process per coefficient in the linear model) are plotted.

main, ylab, ylim, ... 	 high-level plot function parameters.

See Also

mefp

Examples


df1 <- data.frame(y=rnorm(300))
df1[150:300,"y"] <- df1[150:300,"y"]+1
me1 <- mefp(y~1, data=df1[1:50,,drop=FALSE], type="ME", h=1,
              alpha=0.05)
me2 <- monitor(me1, data=df1)

plot(me2)

[Package Contents]