recresid {strucchange}	R Documentation

Recursive Residuals

Description

A generic function for computing the recursive residuals (standardized one step prediction errors) of a linear regression model.

Usage


## Default S3 method:
recresid(x, y, ...)
## S3 method for class 'formula':
recresid(formula, data = list(), ...)
## S3 method for class 'lm':
recresid(x, data = list(), ...)

Arguments

x, y, formula 	 specification of the linear regression model: either by a regressor matrix x and a response variable y, or by a formula or by a fitted object x of class "lm".

data 	 an optional data frame containing the variables in the model. By default the variables are taken from the environment which recresid is called from. Specifying data might also be necessary when applying recresid to a fitted model of class "lm" if this does not contain the regressor matrix and the response.

... 	 currently not used.

Details

Under the usual assumptions for the linear regression model the recdursive residuals are (asymptotically) normal and i.i.d. (see Brown, Durbin, Evans (1975) for details).

Value

A vector containing the recursive residuals.

References

Brown R.L., Durbin J., Evans J.M. (1975), Techniques for testing constancy of regression relationships over time, Journal of the Royal Statistal Society, B, 37, 149-163.

See Also

efp

Examples


x <- rnorm(100)
x[51:100] <- x[51:100] + 2
rr <- recresid(x ~ 1)
plot(cumsum(rr), type = "l")

plot(efp(x ~ 1, type = "Rec-CUSUM"))

[Package Contents]