boundary {strucchange}	R Documentation

Boundary Function for Structural Change Tests

Description

A generic function computing boundaries for structural change tests

Usage

boundary(x, ...)

Arguments

x
an object. Use methods to see which class has a method for boundary.

...
additional arguments affecting the boundary.

Value

an object of class "ts" with the same time properties as the time series in x

See Also

boundary.efp, boundary.mefp, boundary.Fstats

[Package strucchange version 1.5-3 Index]