sctest {strucchange}	R Documentation

Structural Change Tests

Description

Generic function for performing structural change tests.

Usage

sctest(x, ...)

Arguments

x
an object.

...
arguments passed to methods.

Details

sctest is a generic function for performing/extracting structural change tests based on various types of objects. The strucchange package provides various types of methods.

First, structural change tests based on F statistics in linear regression models (Fstats), empirical fluctuation processes in linear regression models (efp), and generalized empirical fluctuation processes in parametric models (gefp) are available in the corresponding sctest methods.

Second, convenience interfaces for carrying out structural change tests in linear regression models and general parametric models are provided in sctest.formula and sctest.default, respectively.

Value

An object of class "htest" containing:

statistic
the test statistic,

p.value
the corresponding p value,

method
a character string with the method used,

data.name
a character string with the data name.

References

Zeileis A., Leisch F., Hornik K., Kleiber C. (2002), strucchange: An R Package for Testing for Structural Change in Linear Regression Models, Journal of Statistical Software, 7(2), 1-38. doi: 10.18637/jss.v007.i02.

Zeileis A. (2006), Implementing a Class of Structural Change Tests: An Econometric Computing Approach. Computational Statistics & Data Analysis, 50, 2987–3008. doi: 10.1016/j.csda.2005.07.001.

See Also

sctest.formula, sctest.default, sctest.Fstats, sctest.efp, sctest.gefp

[Package strucchange version 1.5-3 Index]